requirements.txt
# Core Framework
py4web==1.20241127.1
pydal==20241031.1

# Database Drivers
psycopg2-binary==2.9.10
redis==5.2.1

# Web Server & WSGI
gunicorn==23.0.0
uvicorn[standard]==0.33.0

# Security & Authentication
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==44.0.1

# HTTP & API
requests==2.32.4
httpx==0.28.1
aiohttp==3.12.14

# Configuration
python-decouple==3.8
click==8.1.8

# Monitoring & Metrics
prometheus-client==0.21.1

# Logging
structlog==24.5.0

# Validation
jsonschema==4.23.0
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Development Dependencies (optional)
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.0

# License Server Integration
# (Using requests for HTTP client - already included above)

# IceShelves - LXD/KVM Management
pylxd==2.3.4
libvirt-python==10.10.0
PyYAML==6.0.2
Jinja2==3.1.5